# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NexoShop API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallet/recharge/crypto:
    post:
      operationId: initiateCryptoRecharge
      tags: [wallet]
      summary: Initiate a crypto recharge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CryptoRechargeBody"
      responses:
        "200":
          description: Recharge initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CryptoRechargeResult"
  /wallet/recharge/crypto/verify:
    post:
      operationId: verifyCryptoRecharge
      tags: [wallet]
      summary: Verify crypto payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyCryptoBody"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyCryptoResult"
...
  /wallet/recharge/crypto/pending/{id}:
    delete:
      operationId: cancelPendingCryptoRecharge
      tags: [wallet]
      summary: Cancel a pending crypto recharge
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleOk"
        "404":
          description: Not found
...
  /wallet/recharge/paypal/create:
    post:
      operationId: createPaypalOrder
      tags: [wallet]
      summary: Create a PayPal order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaypalCreateBody"
      responses:
        "200":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaypalCreateResult"
        "503":
          description: PayPal not configured
  /wallet/recharge/paypal/capture:
    post:
      operationId: capturePaypalOrder
      tags: [wallet]
      summary: Capture a PayPal order and credit the wallet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaypalCaptureBody"
      responses:
        "200":
          description: Captured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaypalCaptureResult"
        "503":
          description: PayPal not configured
...
components:
  schemas:
...
    SimpleOk:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
...
    CryptoRechargeBody:
      type: object
      properties:
        amountEur:
          type: number
      required: [amountEur]
    CryptoRechargeResult:
      type: object
      properties:
        sessionId:
          type: integer
        address:
          type: string
        amountLtc:
          type: number
        amountEur:
          type: number
        exchangeRate:
          type: number
        expiresAt:
          type: string
      required: [sessionId, address, amountLtc, amountEur, exchangeRate, expiresAt]
    VerifyCryptoBody:
      type: object
      properties:
        txHash:
          type: string
        amountEur:
          type: number
        sessionId:
          type: integer
          nullable: true
      required: [txHash, amountEur]
    VerifyCryptoResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        newBalance:
          type: ["number", "null"]
      required: [success, message]
...
    PaypalCreateBody:
      type: object
      properties:
        amountEur:
          type: number
      required: [amountEur]
    PaypalCreateResult:
      type: object
      properties:
        orderId:
          type: string
        amountEur:
          type: number
      required: [orderId, amountEur]
    PaypalCaptureBody:
      type: object
      properties:
        orderId:
          type: string
      required: [orderId]
    PaypalCaptureResult:
      type: object
      properties:
        success:
          type: boolean
        newBalance:
          type: number
        amountEur:
          type: number
      required: [success, newBalance, amountEur]
...